fix: fix build post merge
