+++ nextjs/src/app/hub/render/[hostname]/[slug]/page.tsx
@@ -1,9 +1,8 @@
-import { Render } from "@measured/puck/rsc";
 import { getClient } from "@/services/apollo-client";
 import { Metadata, ResolvingMetadata } from "next";
 import { GetPageQuery, GetPageQueryVariables } from "@/__generated__/graphql";
-import { conf } from "@/data/puck/config";
 import { GET_PAGE } from "@/app/hub/render/[hostname]/query";
+import RenderPuck from "../RenderPuck";
 
 type Params = {
   hostname: string
@@ -21,7 +20,7 @@ export default async function Page({ params: { hostname, slug } }: { params: Par
   })
 
   return (
-    <Render config={conf} data={page.data?.hubPageByPath?.puckJsonContent} />
+    <RenderPuck page={page.data?.hubPageByPath?.puckJsonContent} />
   )
 }
