+++ nextjs/src/app/hub/render/[hostname]/map/[[...postcodes]]/page.tsx
@@ -15,6 +15,7 @@ import { GetLocalDataQuery, GetLocalDataQueryVariables } from "@/__generated__/g
 import { SIDEBAR_WIDTH } from "@/components/hub/data";
 import { useRouter } from 'next/navigation';
 import { usePathname, useParams } from 'next/navigation' 
+import { SearchPanel } from './SearchPanel';
 
 type Params = {
   hostname: string
@@ -79,42 +80,6 @@ export default function Page({ params: { hostname, postcodes } }: { params: Para
         </div>
       </JotaiProvider>
     </MapProvider>
-  )
-}
-
-export function SearchPanel ({
-  onSearch,
-  isLoading
-}: {
-  onSearch: (postcode: string) => void,
-  isLoading: boolean
-}) {
-  const [postcode, setPostcode] = useState("")
-
-  return (
-    <>
-      <h1 className='text-2xl font-bold mb-1 leading-tight'>
-        Find out how you can support the climate and nature
-      </h1>
-      <p className='text-sm text-meepGray-500'>
-        Explore our map of Husting events happening all over the uk or input your postcode to see what{"’"}s happening near you. We{"’"}ve had over 300+ events so far.
-      </p>
-      <input
-        type="text"
-        placeholder="Enter your postcode"
-        className='p-4 text-lg w-full rounded-md border border-meepGray-300 mt-4 active:border-green-500'
-        value={postcode}
-        onChange={e => setPostcode(e.target.value)}
-      />
-      <button
-        className='bg-green-500 text-white text-lg font-bold rounded-md w-full p-4 mt-4'
-        // TODO: add postcode validation
-        disabled={!postcode || isLoading}
-        onClick={() => onSearch(postcode)}
-      >
-        {isLoading ? 'Loading...' : 'Search'}
-      </button>
-    </>
   )
 }
